Create `is_gesture_detected` function inside `core/gestures` module

+++ src/core/input.rs
@@ -3,5 +3,5 @@ use crate::enums::KeyboardKey;
 use super::ffi::IsKeyPressed;
 
 pub fn is_key_pressed(key: KeyboardKey) -> bool {
-  return unsafe { IsKeyPressed(key) };
+  return unsafe { IsKeyPressed(key as i32) };
 }
